+++ components/file-upload/index.ts
@@ -1,2 +1,2 @@
+export { default as HeadLessFileUpload } from "./src/HeadLessFileUpload";
 export { default as ReuseFileUpload } from "./src/ReuseFileUpload";
-export { default as HeadlessFileUpload } from "./src/HeadlessFileUpload";