+++ components/file-upload/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @locoworks/reusejs-react-file-upload
 
+## 0.3.0
+
+### Minor Changes
+
+- 50abdd3: Added the setIsDraggedOver prop for the drag and drop feature.
+
+## 0.2.0
+
+### Minor Changes
+
+- b26f1c9: Added the Drag And Drop Feature.
+
 ## 0.1.0
 
 ### Minor Changes